Added back genomes.py for genome defs.
