+++ stdpopsim/species.py
@@ -7,6 +7,7 @@ import logging
 import attr
 import msprime
 
+from . import genomes
 
 logger = logging.getLogger(__name__)
 
@@ -52,7 +53,7 @@ def all_models():
             yield model
 
 
-@attr.s
+@attr.s(frozen=True)
 class Species(object):
     """
     Class representing a species in the catalog.
@@ -132,9 +133,8 @@ class Species(object):
             gm = self.get_genetic_map(genetic_map)
             recomb_map = gm.get_chromosome_map(chrom.name)
 
-        ret = Contig()
-        ret.recombination_map = recomb_map
-        ret.mutation_rate = chrom.mutation_rate
+        ret = genomes.Contig(
+            recombination_map=recomb_map, mutation_rate=chrom.mutation_rate)
         return ret
 
     def get_model(self, id_):
@@ -160,103 +160,3 @@ class Species(object):
             if gm.name == name:
                 return gm
         raise ValueError("Genetic map not found")
-
-
-class Genome(object):
-    """
-    Class representing the genome for a species.
-
-    .. todo:: Define the facilities that this object provides.
-    """
-    def __init__(self, chromosomes):
-        self.chromosomes = chromosomes
-        self.length = 0
-        for chromosome in chromosomes:
-            self.length += chromosome.length
-
-    def __str__(self):
-        s = "Chromosomes:\n"
-        length_sorted = sorted(self.chromosomes, key=lambda x: -x.length)
-        for chrom in length_sorted:
-            s += "\t{}\n".format(chrom)
-        return s
-
-    def get_chromosome(self, name):
-        """
-        Returns the chromosome with the specified name.
-        """
-        for chrom in self.chromosomes:
-            if chrom.name == name:
-                return chrom
-        raise ValueError("Chromosome not found")
-
-    @property
-    def mean_recombination_rate(self):
-        """
-        The length-weighted mean recombination rate across all chromosomes.
-        """
-        mean_recombination_rate = 0
-        for chrom in self.chromosomes:
-            normalized_weight = chrom.length / self.length
-            cont = chrom.recombination_rate * normalized_weight
-            mean_recombination_rate += cont
-        return mean_recombination_rate
-
-    @property
-    def mean_mutation_rate(self):
-        """
-        The length-weighted mean mutation rate across all chromosomes.
-        """
-        mean_mutation_rate = 0
-        for chrom in self.chromosomes:
-            normalized_weight = chrom.length / self.length
-            cont = chrom.mutation_rate * normalized_weight
-            mean_mutation_rate += cont
-        return mean_mutation_rate
-
-
-class Chromosome(object):
-    """
-    Class representing a single chromosome for a species.
-
-    .. todo:: Define the facilities that this object provides.
-    """
-    def __init__(self, name, length, recombination_rate, mutation_rate):
-        self.name = name
-        self.length = length
-        self.recombination_rate = recombination_rate
-        self.mutation_rate = mutation_rate
-
-    def __repr__(self):
-        return (
-            "{{'name': {}, 'length': {}, "
-            "'recombination_rate': {}, "
-            "'mutation_rate': {}}}".format(
-                self.name, self.length, self.recombination_rate,
-                self.mutation_rate))
-
-    def __str__(self):
-        return repr(self)
-
-
-class Contig(object):
-    """
-    Class representing a contiguous region of genome that is to be
-    simulated. This contains all information required to simulate
-    the region, including mutation, recombination rates, etc.
-
-    See :func:`.contig_factory` function for information on how to
-    populate this
-    """
-    def __init__(self):
-        self.recombination_map = None
-        self.mutation_rate = None
-
-    def __str__(self):
-        s = (
-            "Contig(length={:.2G}, recombination_rate={:.2G}, "
-            "mutation_rate={:.2G})").format(
-                self.recombination_map.get_length(),
-                self.recombination_map.mean_recombination_rate,
-                self.mutation_rate)
-        return s
